Changed type of text list's value field from decimal to int. Added Item1 and Item2 for compatibility. KP-9402

+++ src/Tests/WB.Tests.Unit/SharedKernels/SurveyManagement/PreloadedDataVerifierTests/when_verifying_preloaded_data_file_has_orphan_roster_data.cs
@@ -46,7 +46,7 @@ namespace WB.Tests.Unit.SharedKernels.SurveyManagement.PreloadedDataVerifierTest
                 .Returns(preloadedDataByFileTopLevel);
 
             preloadedDataServiceMock.Setup(x => x.GetAvailableIdListForParent(preloadedDataByFileTopLevel, Moq.It.IsAny<ValueVector<Guid>>(), new []{"1"}, files))
-                .Returns((decimal[])null);
+                .Returns((int[])null);
             preloadedDataServiceMock.Setup(x => x.GetColumnIndexByHeaderName(preloadedDataByFileTopLevel, Moq.It.IsAny<string>())).Returns(-1);
             preloadedDataVerifier = CreatePreloadedDataVerifier(questionnaire, preloadedDataServiceMock.Object);
         };
